Add note about client cert validation with skip ssl
Covers the case Alexey found when setting up peering between GCP
environments. Setting up peer replication failed because each env
is configured to request client certificates from Gorouter clients. When
setting up peering, the client cert presented is the autogenerated one
for the container running the service registry instance. The only way
this can be accepted by the target platform is if it trusts the CA that
signed the container cert from the connecting platform.